Organisation-related Functions and Unit Tests
Closes #18
Closes #17

- Function for changing the active Organisation. Essentailly, just making sure the user is a member of the organisation, and returning a new JWT with that OrganisationId if they are.
- Ability to create, update, and delete an Organisation. Requires varying level of credentials, i.e. deleting an Organisation is an administrative action and requires a username & password.

+++ ReactVR_API/ReactVR_API.Core/Functions/OrganisationFunctions.cs
@@ -14,6 +14,7 @@ using ReactVR_API.Core.Security.AccessTokens;
 using ReactVR_API.Core.HelperClasses;
 using System.Linq;
 using System.Security.Claims;
+using ReactVR_API.Core.Security.Login;
 
 namespace ReactVR_API.Core.Functions
 {
@@ -105,6 +106,7 @@ namespace ReactVR_API.Core.Functions
         [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "Organisation/GetOrganisationsForUser")] HttpRequest req, ILogger log)
         {
             log.LogInformation("C# HTTP trigger function(GetOrganisationsForUser) processed a request.");
+
             var accessTokenResult = _tokenProvider.ValidateToken(req);
 
             if (accessTokenResult.Status == AccessTokenStatus.Valid)
@@ -131,21 +133,49 @@ namespace ReactVR_API.Core.Functions
             }
         }
 
-        [FunctionName("UpdateOrganisation")]
-        public async Task<IActionResult> UpdateOrganisation(
-        [HttpTrigger(AuthorizationLevel.Anonymous, "put", Route = "organisation")] HttpRequest req, ILogger log)
+        /// <summary>
+        /// Client will send Id of User who is switching their active Organisation
+        /// They will also send the Id of the Organisation being switched to
+        /// Server can then return a new JWT with the OrganisationId of the target Organisation
+        /// </summary>
+        /// <param name="req"></param>
+        /// <param name="log"></param>
+        /// <returns></returns>
+        [FunctionName("ChangeActiveOrganisation")]
+        public async Task<IActionResult> ChangeActiveOrganisation([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "Organisation/ChangeActiveOrganisation")] HttpRequest req, ILogger log)
         {
-            log.LogInformation("C# HTTP trigger function(UpdateOrganisation) processed a request.");
-
-            string requestBody = await new StreamReader(req.Body).ReadToEndAsync();
-            var organisation = JsonConvert.DeserializeObject<Organisation>(requestBody);
+            log.LogInformation("C# HTTP trigger function(ChangeActiveOrganisation) processed a request.");
 
             try
             {
-                var organisationRepo = new OrganisationRepository();
-                organisationRepo.UpdateOrganisation(organisation);
+                // Validate JWT
+                var accessTokenResult = _tokenProvider.ValidateToken(req);
+                if (accessTokenResult.Status != AccessTokenStatus.Valid)
+                {
+                    return new UnauthorizedResult();
+                }
 
-                return new OkObjectResult($"Updated {organisation.OrganisationName}.");
+                string requestBody = await new StreamReader(req.Body).ReadToEndAsync();
+                // not sure if I should check if this organisation exists first?
+                var targetOrganisation = JsonConvert.DeserializeObject<Organisation>(requestBody);
+
+                Guid userAccountId = new Guid(accessTokenResult.Principal.Claims.First(c => c.Type == "UserAccount").Value);
+                Guid organisationId = new Guid(accessTokenResult.Principal.Claims.First(c => c.Type == "Organisation").Value);
+
+                // Make sure this user is a member of the organisation
+                var organisationMembershipRepository = new OrganisationMembershipRepository();
+                var organisationMembership = organisationMembershipRepository.GetOrganisationMembership(userAccountId, organisationId);
+
+                if (organisationMembership != null)
+                {
+                    // return JWT with UserAccountId
+                    var jwt = _tokenCreator.CreateToken(userAccountId, targetOrganisation.OrganisationId);
+                    return new OkObjectResult(jwt);
+                }
+                else
+                {
+                    return new ConflictObjectResult("User is not a member of this Organisation. Please contact support.");
+                }
             }
             catch (Exception exception)
             {
@@ -153,18 +183,97 @@ namespace ReactVR_API.Core.Functions
             }
         }
 
+        [FunctionName("UpdateOrganisation")]
+        public async Task<IActionResult> UpdateOrganisation(
+        [HttpTrigger(AuthorizationLevel.Anonymous, "put", Route = "organisation")] HttpRequest req, ILogger log)
+        {
+            log.LogInformation("C# HTTP trigger function(UpdateOrganisation) processed a request.");
+
+            var accessTokenResult = _tokenProvider.ValidateToken(req);
+
+            if (accessTokenResult.Status == AccessTokenStatus.Valid)
+            {
+                try
+                {
+                    Guid userAccountId = new Guid(accessTokenResult.Principal.Claims.First(c => c.Type == "UserAccount").Value);
+                    Guid organisationId = new Guid(accessTokenResult.Principal.Claims.First(c => c.Type == "Organisation").Value);
+
+                    string requestBody = await new StreamReader(req.Body).ReadToEndAsync();
+                    var organisationUpdateModel = JsonConvert.DeserializeObject<OrganisationCreateModel>(requestBody);
+
+                    var organisationRepo = new OrganisationRepository();
+                    var organisation = organisationRepo.GetOrganisationById(organisationId);
+
+                    ModelUpdater.CopyProperties(organisationUpdateModel, organisation);
+
+                    bool updated = organisationRepo.UpdateOrganisation(organisation);
+
+                    return new OkObjectResult("Updated");
+                }
+                catch (Exception exception)
+                {
+                    return new BadRequestObjectResult(exception.Message);
+                }
+            }
+            else
+            {
+                return new UnauthorizedResult();
+            }
+        }
+
+        /// <summary>
+        /// Function for deleting an Organisation
+        /// First of all, user must provide a valid JWT which hasn't been tampered with - containing UserAccountId & OrganisationId
+        /// Secondly, since this is an administrative action - username & password must also be provided 
+        /// Thirdly, once validated & logged in - verify that the user is in fact the organisation owner
+        /// </summary>
+        /// <param name="req"></param>
+        /// <param name="log"></param>
+        /// <returns></returns>
         [FunctionName("DeleteOrganisation")]
         public async Task<IActionResult> DeleteOrganisation(
-        [HttpTrigger(AuthorizationLevel.Anonymous, "delete", Route = "organisation/{OrganisationId}")] HttpRequest req, ILogger log, Guid organisationId)
+        [HttpTrigger(AuthorizationLevel.Anonymous, "delete", Route = "Organisation/DeleteOrganisation")] HttpRequest req, ILogger log)
         {
             log.LogInformation("C# HTTP trigger function(DeleteOrganisation) processed a request.");
 
             try
             {
-                var organisationRepo = new OrganisationRepository();
-                organisationRepo.DeleteOrganisation(organisationId);
+                // Validate JWT
+                var accessTokenResult = _tokenProvider.ValidateToken(req);
+                if (accessTokenResult.Status != AccessTokenStatus.Valid)
+                {
+                    return new UnauthorizedResult();
+                }
 
-                return new OkObjectResult($"Deleted {organisationId}");
+                string requestBody = await new StreamReader(req.Body).ReadToEndAsync();
+                var userAccountCreateModel = JsonConvert.DeserializeObject<UserAccountCreateModel>(requestBody);
+
+                // Validate Email/Password
+                var loginManager = new LoginManager();
+                var loginResult = loginManager.AttemptLogin(userAccountCreateModel.EmailAddress, userAccountCreateModel.Password);
+                if (loginResult.Status != LoginStatus.Success)
+                {
+                    return new BadRequestObjectResult(loginResult.FailureReason);
+                }
+
+                Guid userAccountId = new Guid(accessTokenResult.Principal.Claims.First(c => c.Type == "UserAccount").Value);
+                Guid organisationId = new Guid(accessTokenResult.Principal.Claims.First(c => c.Type == "Organisation").Value);
+
+                // Make sure this UserAccount is the Organisation Owner
+                var organisationMembershipRepository = new OrganisationMembershipRepository();
+                var organisationMembership = organisationMembershipRepository.GetOrganisationMembership(userAccountId, organisationId);
+
+                if (organisationMembership.UserTypeId == UserType.OrganisationOwner)
+                {
+                    var organisationRepo = new OrganisationRepository();
+                    bool deleted = organisationRepo.DeleteOrganisation(organisationId);
+
+                    return new OkObjectResult(deleted);
+                }
+                else
+                {
+                    return new UnauthorizedResult();
+                }
             }
             catch (Exception exception)
             {
